+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 ### 新特性
 * 【core】         增加RejectPolicy线程池线程拒绝策略枚举
 * 【core】         DateUtil增加isSame方法
+* 【core】         FileUtil.getAbsolutePath方法在获取不到ClassPath情况下返回原路径
 
 ### Bug修复
 * 【core】         修复RuntimeUtil.getResultLines未关闭Process问题（pr#164@Github）
+* 【core】         修复ClassPathResource在jar运行模式下的空指针问题
 
 -------------------------------------------------------------------------------------------------------------